requirements.txt
openai==0.28
playwright==1.51.0
pymongo==4.11.3
numpy==2.2.4
langgraph==0.3.21
psycopg2-binary==2.9.10
python-dotenv==1.1.0
langchain-community==0.3.20
langchain-postgres==0.0.13
psycopg-binary==3.2.6
black==25.1.0
requests==2.32.3
fastapi==0.115.12
uvicorn==0.34.0